requirements.txt
cloudpickle==3.0.0
Farama-Notifications==0.0.4
filelock==3.13.1
fsspec==2024.2.0
gymnasium==0.29.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
pillow==10.2.0
sympy==1.12
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
wandb~=0.16.4
minigrid~=2.3.1
PyYAML~=6.0.1